@@ -1,5 +1,6 @@
 import { checkInternetConnection } from 'react-native-offline';
-import * as Expo from 'expo';
+import { Asset } from 'expo-asset';
+import * as FileSystem from 'expo-file-system';
 import Database from './Database';
 import * as store from 'react-native-simple-store';
 
@@ -8,10 +9,10 @@ const NEW_HASH = '2';
 
 describe('Database setup', () => {
   test('if file exists and hash hasnt changed, then db is available', async () => {
-    Expo.FileSystem.getInfoAsync = jest.fn(() => {
+    FileSystem.getInfoAsync = jest.fn(() => {
       return { exists: true };
     });
-    Expo.Asset.fromModule = jest.fn(() => {
+    Asset.fromModule = jest.fn(() => {
       return { hash: null };
     });
     // Had trouble mocking this, so by default it returns null
@@ -20,10 +21,10 @@ describe('Database setup', () => {
     expect(available).toBe(true);
   });
   test('if file does not exist, then db is NOT available', async () => {
-    Expo.FileSystem.getInfoAsync = jest.fn(() => {
+    FileSystem.getInfoAsync = jest.fn(() => {
       return { exists: false };
     });
-    Expo.Asset.fromModule = jest.fn(() => {
+    Asset.fromModule = jest.fn(() => {
       return { hash: EXISTING_HASH };
     });
     store.get = jest.fn(() => EXISTING_HASH);
@@ -31,10 +32,10 @@ describe('Database setup', () => {
     expect(available).toBe(false);
   });
   test('if hash has changed, then db is NOT available', async () => {
-    Expo.FileSystem.getInfoAsync = jest.fn(() => {
+    FileSystem.getInfoAsync = jest.fn(() => {
       return { exists: false };
     });
-    Expo.Asset.fromModule = jest.fn(() => {
+    Asset.fromModule = jest.fn(() => {
       return { hash: NEW_HASH };
     });
     store.get = jest.fn(() => EXISTING_HASH);
